// typescript-projects/utility-backups/src20250428workingPlanModel/interface/types.ts
export enum UtilityMerchant {
  // Mobile merchants
  Mtn = "mtn",
  Airtel = "airtel",
  Glo = "glo",
  NineMobile = "9mobile",
  Spectranet = "spectranet",
  // Electricity merchants
  IkejaElectric = "ikejaElectric",
  AbujaElectric = "abujaElectric",
  EkoElectric = "ekoElectric",
  KadunaElectric = "kadunaElectric",
  AbaElectric = "abaElectric",
  // Cable TV merchants
  Dstv = "dstv",
  Gotv = "gotv",
  Startimes = "startimes",
  // Bet merchants
  Betking = "betking",
  Bet9ja = "bet9ja",
  Betway = "betway",
  Sportybet = "sportybet",
  Nairabet = "nairabet",
}
// TODO: merchantService -> product
export enum UtilityProduct {
  Airtime = "airtime",
  Prepaid = "prepaid",
  Postpaid = "postpaid",
  Plan = "plan",
  Default = "default",
}

// TODO: update
export type TPlanData = {
  providerId: ProviderId;
  planCode: string;
  amount: number;
};

export type TPaymentData = TUtilityMeta &
  (
    | {
        // Require plan data as well, if service is of 'plan' type...
        productId: UtilityProduct.Plan;
        planData: TPlanData;
      }
    | {
        // ...otherwise, it should not be present.
        productId: Exclude<UtilityProduct, "plan">;
        planData?: undefined;
      }
  );

/** Data expected to come into the payment-processing endpoint service class */
export type TProcessData = {
  paymentData: TPaymentData;
  providerId: ProviderId;
  uid: string;
  appTxHash: string;
  nairaAdminDocPath: string;
  txPath: string;
};

/** Data expected to come into the status-processing endpoint service class */
export type TStatusData = {
  providerId: ProviderId;
  queryStatus: AppTxStatus;
  reference?: string;
  txPath?: string;
};

export interface IUtilityProvider {
  /** Maps merchant product IDs recognized by our app to those of a specific provider. */
  appToProviderProductMap: {
    [K in UtilityMerchant]: { [K2 in UtilityProduct]?: string };
  };

  /** Maps merchant IDs recognized by our app to those of a specific provider. */
  appToProviderMerchantMap: { [K in UtilityMerchant]: string };

  /**
   * Determines if a provider supports a merchant service/product. Only used for non-plan services like airtime.
   * For the 'plan' merchant service, provider ID is expected in fetched plan data.
   * @param service - the merchant service ID recognized by our app.
   * @returns true if specified service/product payment is supported
   */
  supports(merchant: UtilityMerchant, service: UtilityProduct): boolean;

  /**
   * Make utility payment for airtime.
   * @param paymentData - all data that'll be needed to make the utility payment.
   * @param txPath - full path to doc containing request data; used here for generating request reference.
   * @returns data to help with further processing of the request at the business logic level.
   */
  [UtilityProduct.Airtime](
    paymentData: TPaymentData,
    txPath: string
  ): Promise<IProviderPaymentResult>;

  /**
   * Make utility payment for a merchant's default service (variable pricing).
   * @param paymentData - all data that'll be needed to make the utility payment.
   * @param txPath - full path to doc containing request data; used here for generating tx reference.
   * @returns data to help with further processing of the request at the business logic level.
   */
  [UtilityProduct.Default](
    paymentData: TPaymentData,
    txPath: string
  ): Promise<IProviderPaymentResult>;

  /**
   * Make utility payment for merchant services grouped into bundles with fixed prices.
   * @param paymentData - all data that'll be needed to make the utility payment.
   * @param txPath - full path to doc containing request data; used here for generating tx reference.
   * @returns data to help with further processing of the request at the business logic level.
   */
  [UtilityProduct.Plan](
    paymentData: TPaymentData,
    txPath: string
  ): Promise<IProviderPaymentResult>;

  /**
   * Make utility payment for electricity postpaid service.
   * @param paymentData - all data that'll be needed to make the utility payment.
   * @param txPath - full path to doc containing request data; used here for generating tx reference.
   * @returns data to help with further processing of the request at the business logic level.
   */
  [UtilityProduct.Postpaid](
    paymentData: TPaymentData,
    txPath: string
  ): Promise<IProviderPaymentResult>;

  /**
   * Make utility payment for electricity prepaid service.
   * @param paymentData - all data that'll be needed to make the utility payment.
   * @param txPath - full path to doc containing request data; used here for generating tx reference.
   * @returns data to help with further processing of the request at the business logic level.
   */
  [UtilityProduct.Prepaid](
    paymentData: TPaymentData,
    txPath: string
  ): Promise<IProviderPaymentResult>;

  /** Returns the current transaction status of the bill payment request */
  checkStatus(
    statusData: TStatusData,
    reqBody: Record<string, unknown>
  ): Promise<TCheckStatusResult>;

  /**
   * Derive a request reference given a transaction path.
   * @param txPath - full path to transaction doc (users_transactions_withdrawal).
   * @returns reference used to make the payment request to the provider.
   */
  txpathToTxref(txPath: string): string;
  /**
   * Derive a transaction path given a request reference.
   * @param txRef - reference used to make the payment request to the provider.
   * @returns full path to transaction doc (users_transactions_withdrawal).
   */
  txrefToTxpath(txRef: string): string;
}

/** Unique identifier for each of our providers. E.g. vtpass, remita, paga */
export enum ProviderId {
  ProviderX = "providerX", // TODO: remove; test provider
}

export enum ProviderTxStatus {
  Pending = "pending",
  Successfull = "successful",
  Failed = "failed",
}
export enum AppTxStatus {
  Pending = "pending",
  Approved = "approved",
  Rejected = "rejected",
  Mempool = "mempool",
}

export interface IUtilityStatusTxResult {
  statusDocsWritten: boolean;
}

// TODO: add other metadata from the payment request to save in DB
export interface IProviderPaymentResult {
  initialStatus: ProviderTxStatus;
  /** How status update is retrieved; push-to-app (webhook) or pull-from-provider (requery) */
  statusUpdateMethods: { isWebhook: boolean; isRequery: boolean };
  /** Unique (per utility request to provider) reference sent in utility request */
  txRef: string;
}

/** Data for getting document refs of docs to be written/updated during utility processing. */
export type TStatusDocsAccessData = {
  uid: string;
  /** hash of naira withdrawal tx, which also serves as tx doc ID */
  appTxHash: string;
  /** full path to the related naira_admin doc */
  nairaAdminDocPath: string;
};

/** Result of a provider transaction status check and/or parsing */
export type TCheckStatusResult = { status: AppTxStatus; txPath: string };

// ------------------- MODEL -------------------

/** Contains data about the utility payment request */
export type TUtilityMeta = {
  /** Unique identifier for the customer, at the merchant */
  customerId: string;
  merchantId: UtilityMerchant;
  productId: UtilityProduct;
  planId?: string;
  amount?: number;
} & (
  | {
      // Require planId if service is of 'plan' type...
      productId: UtilityProduct.Plan;
      planId: string;
      amount?: undefined;
    }
  | {
      // ...else, require amount instead
      productId: Exclude<UtilityProduct, "plan">;
      planId?: undefined;
      amount: number;
    }
);